Create guess.py
simple guessing game. The user gets a limited amount of time to guess a randomly generated number 1-10 and if the play either guesses outside of that range or guesses an invalid number, the system will not take away any tries. Only when the guess is wrong and within range will the code take away 1 try from the user